strip open3d_ prefix
