+++ src/Python/visualization/visualization.cpp
@@ -24,6 +24,16 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d_io.h"
+#include "visualization.h"
 
-void pybind_io(py::module &m) {}
+void pybind_visualization(py::module &m) {
+    py::module m_visualization = m.def_submodule("visualization");
+    pybind_renderoption(m_visualization);
+    pybind_viewcontrol(m_visualization);
+    pybind_visualizer(m_visualization);
+    pybind_utility(m_visualization);
+    pybind_renderoption_method(m_visualization);
+    pybind_viewcontrol_method(m_visualization);
+    pybind_visualizer_method(m_visualization);
+    pybind_utility_methods(m_visualization);
+}
